[gui, games] Remove `thprac_` filename prefix
