+++ thprac/src/thprac/alcostg.cpp
