+++ thprac/src/thprac/gui_impl_dx8.h
