+++ thprac/src/thprac/gui_impl_dx9.h
