+++ thprac/src/thprac/gui_impl_win32.h
