+++ thprac/src/thprac/gui_locale.h
