+++ thprac/src/thprac/main.cpp
@@ -4,7 +4,7 @@
 #include "thprac_launcher_main.h"
 #include "thprac_launcher_cfg.h"
 #include "attach.h"
-#include "thprac_gui_locale.h"
+#include "gui_locale.h"
 #include "thprac_hook.h"
 #include <Windows.h>
 #include <psapi.h>
