+++ thprac/src/thprac/th06.cpp
