+++ thprac/src/thprac/th07.cpp
