+++ thprac/src/thprac/th12.cpp
