+++ thprac/src/thprac/thprac_launcher_games.cpp
@@ -1,7 +1,7 @@
 ﻿#define NOMINMAX
 
 #include "thprac_launcher_games.h"
-#include "thprac_gui_locale.h"
+#include "gui_locale.h"
 #include "thprac_launcher_cfg.h"
 #include "thprac_launcher_games_def.h"
 #include "thprac_launcher_main.h"
