+++ thprac/src/thprac/thprac_launcher_games.h
@@ -1,6 +1,6 @@
 ﻿#pragma once
 #include "utils/utils.h"
-#include "thprac_gui_locale.h"
+#include "gui_locale.h"
 #include "thprac_launcher_games_def.h"
 #include <cstdint>
 #pragma warning(push)
