+++ thprac/src/thprac/thprac_launcher_links.cpp
@@ -1,7 +1,7 @@
 ﻿#include "thprac_launcher_cfg.h"
 #include "thprac_launcher_utils.h"
 #include "thprac_launcher_wnd.h"
-#include "thprac_gui_locale.h"
+#include "gui_locale.h"
 #include "common.h"
 #include "utils/utils.h"
 #include <functional>
