+++ thprac/src/thprac/thprac_launcher_main.cpp
@@ -1,5 +1,5 @@
 ﻿#include "thprac_launcher_main.h"
-#include "thprac_gui_locale.h"
+#include "gui_locale.h"
 #include "thprac_launcher_cfg.h"
 #include "thprac_launcher_games.h"
 #include "thprac_launcher_links.h"
