+++ thprac/src/thprac/thprac_launcher_tools.cpp
@@ -3,7 +3,7 @@
 #include "thprac_launcher_games.h"
 #include "thprac_launcher_main.h"
 #include "attach.h"
-#include "thprac_gui_locale.h"
+#include "gui_locale.h"
 #include "common.h"
 #include "utils/utils.h"
 #include <functional>
