+++ thprac/src/thprac/thprac_launcher_wnd.cpp
@@ -1,8 +1,8 @@
 ﻿#include "thprac_launcher_wnd.h"
 #include "imgui.h"
-#include "thprac_gui_impl_dx9.h"
-#include "thprac_gui_impl_win32.h"
-#include "thprac_gui_locale.h"
+#include "gui_impl_dx9.h"
+#include "gui_impl_win32.h"
+#include "gui_locale.h"
 #include "thprac_launcher_utils.h"
 #include "common.h"
 #include "..\..\resource.h"
