+++ thprac/thprac.vcxproj
@@ -48,13 +48,13 @@
     <ClInclude Include="src\3rdParties\MinHook\src\hde\table64.h" />
     <ClInclude Include="src\3rdParties\MinHook\src\trampoline.h" />
     <ClInclude Include="src\thprac\thprac_game_data.h" />
-    <ClInclude Include="src\thprac\thprac_gui_impl_dx8.h" />
-    <ClInclude Include="src\thprac\thprac_gui_components.h" />
+    <ClInclude Include="src\thprac\gui_impl_dx8.h" />
+    <ClInclude Include="src\thprac\gui_components.h" />
     <ClInclude Include="src\thprac\thprac_hook.h" />
-    <ClInclude Include="src\thprac\thprac_gui_impl_dx9.h" />
-    <ClInclude Include="src\thprac\thprac_gui_impl_win32.h" />
+    <ClInclude Include="src\thprac\gui_impl_dx9.h" />
+    <ClInclude Include="src\thprac\gui_impl_win32.h" />
     <ClInclude Include="src\thprac\thprac_init.h" />
-    <ClInclude Include="src\thprac\thprac_gui_input.h" />
+    <ClInclude Include="src\thprac\gui_input.h" />
     <ClInclude Include="src\thprac\thprac_launcher_cfg.h" />
     <ClInclude Include="src\thprac\thprac_launcher_games.h" />
     <ClInclude Include="src\thprac\thprac_launcher_games_def.h" />
@@ -65,7 +65,7 @@
     <ClInclude Include="src\thprac\thprac_launcher_wnd.h" />
     <ClInclude Include="src\thprac\thprac_licence.h" />
     <ClInclude Include="src\thprac\thprac_load_exe.h" />
-    <ClInclude Include="src\thprac\thprac_gui_locale.h" />
+    <ClInclude Include="src\thprac\gui_locale.h" />
     <ClInclude Include="src\thprac\thprac_locale_def.h" />
     <ClInclude Include="src\thprac\attach.h" />
     <ClInclude Include="src\thprac\common.h" />
@@ -96,14 +96,14 @@
     <ClCompile Include="src\3rdParties\MinHook\src\trampoline.c" />
     <ClCompile Include="src\3rdParties\MinHook\src\trampoline_new.c" />
     <ClCompile Include="src\thprac\main.cpp" />
-    <ClCompile Include="src\thprac\thprac_alcostg.cpp" />
-    <ClCompile Include="src\thprac\thprac_gui_impl_dx8.cpp" />
-    <ClCompile Include="src\thprac\thprac_gui_components.cpp" />
+    <ClCompile Include="src\thprac\alcostg.cpp" />
+    <ClCompile Include="src\thprac\gui_impl_dx8.cpp" />
+    <ClCompile Include="src\thprac\gui_components.cpp" />
     <ClCompile Include="src\thprac\thprac_hook.cpp" />
-    <ClCompile Include="src\thprac\thprac_gui_impl_dx9.cpp" />
-    <ClCompile Include="src\thprac\thprac_gui_impl_win32.cpp" />
+    <ClCompile Include="src\thprac\gui_impl_dx9.cpp" />
+    <ClCompile Include="src\thprac\gui_impl_win32.cpp" />
     <ClCompile Include="src\thprac\thprac_init.cpp" />
-    <ClCompile Include="src\thprac\thprac_gui_input.cpp" />
+    <ClCompile Include="src\thprac\gui_input.cpp" />
     <ClCompile Include="src\thprac\thprac_launcher_cfg.cpp" />
     <ClCompile Include="src\thprac\thprac_launcher_games.cpp" />
     <ClCompile Include="src\thprac\thprac_launcher_links.cpp" />
@@ -112,28 +112,28 @@
     <ClCompile Include="src\thprac\thprac_launcher_wnd.cpp" />
     <ClCompile Include="src\thprac\thprac_licence.cpp" />
     <ClCompile Include="src\thprac\thprac_load_exe.cpp" />
-    <ClCompile Include="src\thprac\thprac_gui_locale.cpp" />
+    <ClCompile Include="src\thprac\gui_locale.cpp" />
     <ClCompile Include="src\thprac\thprac_locale_def.cpp" />
     <ClCompile Include="src\thprac\attach.cpp" />
-    <ClCompile Include="src\thprac\thprac_th06.cpp" />
-    <ClCompile Include="src\thprac\thprac_th07.cpp" />
-    <ClCompile Include="src\thprac\thprac_th08.cpp" />
-    <ClCompile Include="src\thprac\thprac_th09.cpp" />
-    <ClCompile Include="src\thprac\thprac_th095.cpp" />
-    <ClCompile Include="src\thprac\thprac_th10.cpp" />
-    <ClCompile Include="src\thprac\thprac_th11.cpp" />
-    <ClCompile Include="src\thprac\thprac_th12.cpp" />
-    <ClCompile Include="src\thprac\thprac_th125.cpp" />
-    <ClCompile Include="src\thprac\thprac_th128.cpp" />
-    <ClCompile Include="src\thprac\thprac_th13.cpp" />
-    <ClCompile Include="src\thprac\thprac_th14.cpp" />
-    <ClCompile Include="src\thprac\thprac_th143.cpp" />
-    <ClCompile Include="src\thprac\thprac_th15.cpp" />
-    <ClCompile Include="src\thprac\thprac_th16.cpp" />
-    <ClCompile Include="src\thprac\thprac_th165.cpp" />
-    <ClCompile Include="src\thprac\thprac_th17.cpp" />
-    <ClCompile Include="src\thprac\thprac_th18.cpp" />
-    <ClCompile Include="src\thprac\thprac_th185.cpp" />
+    <ClCompile Include="src\thprac\th06.cpp" />
+    <ClCompile Include="src\thprac\th07.cpp" />
+    <ClCompile Include="src\thprac\th08.cpp" />
+    <ClCompile Include="src\thprac\th09.cpp" />
+    <ClCompile Include="src\thprac\th095.cpp" />
+    <ClCompile Include="src\thprac\th10.cpp" />
+    <ClCompile Include="src\thprac\th11.cpp" />
+    <ClCompile Include="src\thprac\th12.cpp" />
+    <ClCompile Include="src\thprac\th125.cpp" />
+    <ClCompile Include="src\thprac\th128.cpp" />
+    <ClCompile Include="src\thprac\th13.cpp" />
+    <ClCompile Include="src\thprac\th14.cpp" />
+    <ClCompile Include="src\thprac\th143.cpp" />
+    <ClCompile Include="src\thprac\th15.cpp" />
+    <ClCompile Include="src\thprac\th16.cpp" />
+    <ClCompile Include="src\thprac\th165.cpp" />
+    <ClCompile Include="src\thprac\th17.cpp" />
+    <ClCompile Include="src\thprac\th18.cpp" />
+    <ClCompile Include="src\thprac\th185.cpp" />
     <ClCompile Include="src\thprac\common.cpp" />
     <ClCompile Include="src\thprac\thprac_version.cpp" />
   </ItemGroup>
